requirements.txt
playwright==1.40.0
selenium==4.15.2
beautifulsoup4==4.12.2
scrapy==2.11.0
requests==2.31.0
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
fake-useragent==1.4.0
undetected-chromedriver==3.5.4
lxml==4.9.3
aiohttp==3.9.1
asyncio==3.4.3
pydantic==2.5.0
loguru==0.7.2
retrying==1.3.4